Restructure the single module into a package
This restructuring is overdone: there are far too many small modules.

It is mainly to showcase how to divide a package into subpackages and
modules, as well as use relative imports, the use of `__all__.py`, and
the use of `__init__.py` (which export selected subpackage attributes
directly) and helper "core" modules.